requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anthropic==0.73.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
boolean.py==5.0
build==1.3.0
CacheControl==0.14.4
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.7
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
cyclonedx-python-lib==11.6.0
Cython==3.2.2
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
EbookLib==0.20
ecdsa==0.19.1
edge-tts==7.2.3
email-validator==2.3.0
fastapi==0.115.0
filelock==3.20.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-tools==1.76.0
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.20.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
img2pdf==0.6.3
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==0.3.27
langchain-classic==1.0.0
langchain-community==0.3.31
langchain-core==0.3.80
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.5.0
license-expression==30.4.4
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.2
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.14.0
networkx==3.6.1
nltk==3.9.2
numpy==1.26.4
oauthlib==3.3.1
ocrmypdf==16.11.1
ollama==0.6.0
onnxruntime==1.23.2
openai==2.13.0
opencv-python==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.10.7
ormsgpack==1.12.1
overrides==7.7.0
packageurl-python==0.17.6
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pi_heif==1.1.1
pikepdf==10.0.1
Pillow==10.1.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
piper-tts==1.3.0
platformdirs==4.5.1
playwright==1.57.0
pluggy==1.6.0
portalocker==2.10.1
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.2
psutil==5.9.6
psycopg2-binary==2.9.11
py-serializable==2.1.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.12.0
pydantic_core==2.14.6
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.6
pyparsing==3.3.1
pypdf==6.2.0
PyPDF2==3.0.1
pypdfium2==5.1.0
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.18
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.7.0
qrcode==7.4.2
referencing==0.37.0
regex==2025.11.3
reportlab==4.0.7
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.3.2
scipy==1.11.4
sendgrid==6.12.5
sentence-transformers==2.2.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==0.38.6
streamlit==1.52.1
stripe==7.8.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.3.0
tomli_w==1.2.0
torch==2.1.0
torchaudio==2.1.2+cu118
torchvision==0.16.0
tornado==6.5.4
tqdm==4.66.1
transformers==4.35.2
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.25.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
